add run.sh for calcite_crystal case
