+++ applications/solvers/dbs_pimpleFoamCi/Make/files
@@ -1,0 +1,3 @@
+dbs_pimpleFoamCi.C
+
+EXE = $(FOAM_USER_APPBIN)/dbs_pimpleFoamCi
