+++ applications/solvers/dbs_pimpleFoamCi_raw/Make/files
@@ -1,0 +1,3 @@
+dbs_pimpleFoamCi_v4.C
+
+EXE = $(FOAM_USER_APPBIN)/dbs_pimpleFoamCi_v4
