+++ applications/solvers/dbs_pimpleFoamCi_raw/SRFPimpleFoam/Make/files
@@ -1,0 +1,3 @@
+SRFPimpleFoam.C
+
+EXE = $(FOAM_APPBIN)/SRFPimpleFoam
