+++ applications/solvers/dbs_pimpleFoamCi_raw/correctPhi.H
@@ -1,0 +1,12 @@
+CorrectPhi
+(
+    U,
+    phi,
+    p,
+    dimensionedScalar("rAUf", dimTime, 1),
+    geometricZeroField(),
+    pimple,
+    true
+);
+
+#include "continuityErrs.H"
