--- applications/solvers/dbs_pimpleFoamCi_v4/Make/files
@@ -1,3 +1,0 @@
-dbs_pimpleFoamCi_v4.C
-
-EXE = $(FOAM_USER_APPBIN)/dbs_pimpleFoamCi_v4
