--- applications/solvers/dbs_pimpleFoamCi_v4/SRFPimpleFoam/Make/files
@@ -1,3 +1,0 @@
-SRFPimpleFoam.C
-
-EXE = $(FOAM_APPBIN)/SRFPimpleFoam
