--- applications/solvers/dbs_pimpleFoamCi_v4/SRFPimpleFoam/UrelEqn.H
@@ -1,19 +1,0 @@
-    // Relative momentum predictor
-    tmp<fvVectorMatrix> tUrelEqn
-    (
-        fvm::ddt(Urel)
-      + fvm::div(phi, Urel)
-      + turbulence->divDevReff(Urel)
-      + SRF->Su()
-     ==
-        fvOptions(Urel)
-    );
-    fvVectorMatrix& UrelEqn = tUrelEqn.ref();
-
-    UrelEqn.relax();
-
-    fvOptions.constrain(UrelEqn);
-
-    solve(UrelEqn == -fvc::grad(p));
-
-    fvOptions.correct(Urel);
