--- applications/solvers/dbs_pimpleFoamCi_v4/correctPhi.H
@@ -1,12 +1,0 @@
-CorrectPhi
-(
-    U,
-    phi,
-    p,
-    dimensionedScalar("rAUf", dimTime, 1),
-    geometricZeroField(),
-    pimple,
-    true
-);
-
-#include "continuityErrs.H"
